Refactor attributes
- Move types into _types package
- Absorb aliases into _types/attribute.py module
- Namepace attributes/attribute_type as attributes/type
- Improve comments explaining forward references + recursive dependencies
