+++ docs/beta/attributes.md
@@ -1,5 +1,5 @@
 # Attributes
 
   * [Attribute](/beta/attributes/attribute)
-  * [Attribute Type](/beta/attributes/attribute_type)
+  * [Attribute Type](/beta/attributes/type)
   * [SubAttribute](/beta/attributes/subattribute)
