+++ tamr_client/_types/__init__.py
@@ -1,2 +1,19 @@
+from .attribute import (
+    Array,
+    Attribute,
+    AttributeType,
+    BOOLEAN,
+    ComplexType,
+    DEFAULT,
+    DOUBLE,
+    GEOSPATIAL,
+    INT,
+    LONG,
+    Map,
+    PrimitiveType,
+    Record,
+    STRING,
+    SubAttribute,
+)
 from .json import JsonDict
 from .url import URL
