+++ tamr_client/attributes/__init__.py
@@ -1,2 +1,1 @@
-# This __init__.py file is necessary for `mypy --package`
-# See https://github.com/python/mypy/issues/5759
+from . import type
